+++ Sources/SwiftUIKit/Progress/LinearProgressBar.swift
@@ -41,34 +41,18 @@ import SwiftUI
  */
 public struct LinearProgressBar: View {
 
-    /**
-     Create a linear progress bar.
-
-     - Parameters:
-       - progress: The percentage progress, from `0` to `1`.
-       - style: The style to apply, by default `.standard`.
-       - animation: The animation to use when the value changes, by default `.progressBounce`.
-     */
-    public init(
-        progress: Double,
-        style: Style = .standard,
-        animation: Animation = .progressBounce
-    ) {
+    /// Create a linear progress bar.
+    ///
+    /// - Parameters:
+    ///   - progress: The percentage progress, from `0` to `1`.
+    public init(progress: Double) {
         self.progress = progress
-        self.style = style
-        self.animation = animation
     }
 
-
-    /// The animation to use when the value changes.
-    public let animation: Animation
-
-    /// The percentage progress.
-    public var progress: Double
-
-    /// The style to apply.
-    public let style: LinearProgressBar.Style
-
+    private let progress: Double
+    
+    @Environment(\.linearProgressBarStyle)
+    private var style
 
     public var body: some View {
         GeometryReader { geo in
@@ -79,7 +63,7 @@ public struct LinearProgressBar: View {
                 .clipShape(Capsule())
         }
         .frame(height: style.height)
-        .animation(animation, value: progress)
+        .animation(style.animation, value: progress)
         .accessibilityElement()
     }
 }
@@ -107,122 +91,6 @@ private extension LinearProgressBar {
 }
 
 
-// MARK: - Style
-
-public extension LinearProgressBar {
-
-    /**
-     This style can be used to style ``LinearProgressBar``.
-     */
-    struct Style {
-
-        /**
-         Create a linear progress bar style.
-
-         When creating a style with a background color, note
-         that the color will be placed below the material to
-         create a nice frosted effect. If you want the plain
-         color, just set `backgroundMaterialOpacity` to 0.
-
-         - Parameters:
-           - height: The total height of the bar, by default `10`.
-           - backgroundMaterial: The background material, by default `.ultraThin`.
-           - backgroundMaterialOpacity: The background material opacity, by default `1`.
-           - backgroundColor: The background color, by default `.clear`.
-           - barColor: The color of the progress bar, by default `.accentColor`.
-           - barPadding: The padding between the progress bar and the background, by default `0`.
-           - barShadow: The shadow style to apply to the progress bar, by default `.progressbar`.
-         */
-        public init(
-            height: Double = 10,
-            backgroundMaterial: Material = .ultraThin,
-            backgroundMaterialOpacity: Double = 1,
-            backgroundColor: Color = .clear,
-            barColor: Color = .accentColor,
-            barPadding: Double = 0,
-            barShadow: ViewShadowStyle = .progressBar
-        ) {
-            self.height = height
-            self.backgroundColor = backgroundColor
-            self.backgroundMaterial = backgroundMaterial
-            self.backgroundMaterialOpacity = backgroundMaterialOpacity
-            self.barColor = barColor
-            self.barPadding = barPadding
-            self.barShadow = barShadow
-        }
-
-        public typealias ViewModifier = (AnyView) -> AnyView
-
-
-        /// The total height of the bar.
-        public var height: Double
-
-        /// The background material.
-        public var backgroundMaterial: Material
-
-        /// The background material opacity.
-        public var backgroundMaterialOpacity: Double
-
-        /// The background color.
-        public var backgroundColor: Color
-
-        /// The color of the progress bar.
-        public var barColor: Color
-
-        /// The padding between the progress bar and the background.
-        public var barPadding: Double
-
-        /// The shadow style to apply to the progress bar.
-        public var barShadow: ViewShadowStyle
-    }
-}
-
-
-// MARK: - Styles
-
-public extension LinearProgressBar.Style {
-
-    /// The standard linear progress bar style.
-    static var standard = LinearProgressBar.Style()
-
-    /// A style that adds padding to the bar.
-    static var padding = LinearProgressBar.Style(
-        barPadding: 2
-    )
-
-    /// A style that makes the bar tall and adds padding.
-    static var tallPadding = LinearProgressBar.Style(
-        height: 16,
-        barPadding: 3
-    )
-}
-
-
-// MARK: - Shadows
-
-public extension ViewShadowStyle {
-
-    /**
-     This shadow style can be used with ``LinearProgressBar``.
-     */
-    static let progressBar = ViewShadowStyle(
-        color: .black.opacity(0.4),
-        radius: 2,
-        x: 1)
-}
-
-
-// MARK: - Animations
-
-public extension Animation {
-
-    /**
-     This animation can be used with ``LinearProgressBar``.
-     */
-    static var progressBounce = Animation.interpolatingSpring(mass: 1, stiffness: 100, damping: 12, initialVelocity: 0)
-}
-
-
 // MARK: - Previews
 
 #Preview {
@@ -236,7 +104,8 @@ public extension Animation {
 
         var body: some View {
             VStack(spacing: 30) {
-                LinearProgressBar(progress: progress, style: style)
+                LinearProgressBar(progress: progress)
+                    .linearProgressBarStyle(style)
                 #if os(iOS) || os(macOS) || os(watchOS)
                     .onTapGesture {
                         progress += 0.1
@@ -264,7 +133,8 @@ public extension Animation {
             Preview(style: .swedish)
             Preview(style: .swedishFrosted)
                 .shadow(.elevated)
-        }.padding(50)
+        }
+        .padding(50)
     }
 }
 
@@ -280,6 +150,7 @@ private extension LinearProgressBar.Style {
 
     static var swedishFrosted: LinearProgressBar.Style = {
         var style = LinearProgressBar.Style.tallPadding
+        style.animation = .interpolatingSpring()
         style.backgroundColor = .blue
         style.barColor = .yellow
         return style